pyproject.toml
[project]
name = "searchtool-gtk"
version = "2.0.0"
description = "A generic GTK search tool and launcher"
license = "Unlicense"
readme = "README.md"
requires-python = ">=3.13,<4.0"
authors = [
  { "name" = "Ianis Vasilev", "email" = "ianis@ivasilev.net" }
]

dependencies = [
  "PyGObject (>=3.54.5)",
  "pyxdg (>=0.28)",
  "pydantic (>=2.12.3)",
  "pyicu (>=2.15.3,<3.0.0)",
]

[project.urls]
Repository = "https://github.com/v--/searchtool-gtk.git"

[project.scripts]
searchtool-gtk-server = "searchtool_gtk:entry_point"

[dependency-groups]
dev = [
  "PyGObject-stubs (>=2.13.0)",
  "mypy (>=1.18.2)",
  "ruff (>=0.14.2)",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = ["UP", "I"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
packages = ["searchtool_gtk"]

[[tool.mypy.overrides]]
module = [
  "gi.*",
  "icu.*",
  "xdg.*"
]
ignore_missing_imports = true
